feat: getChipColorByStatus
