+++ src/utils/index.js
@@ -4,3 +4,4 @@ export { default as tableFiller } from './tableFiller';
 export { default as InternshipData } from './internshipData';
 export { default as tableFillerAllUsers } from './tableFillerAllUsers';
 export { default as tableFillerCandidates } from './tableFillerCandidates';
+export { default as getChipColorByStatus } from './getChipColorByStatus';
